Improve err msg when source is missing in `BinFile::install_bin` (#301)
* Improve err msg in `helpers::atomic_install`


* Check for missing `source` in `install_bin`


* Fix `install_bin`: Use `try_exists` instead of `exists`


* Do not test `cargo-deb` in `ci-scripts/tests.sh`

Due to the fact that it uses gnu sparse extension which is not supported
by crate `tar`.

+++ src/bins.rs
@@ -79,7 +79,10 @@ impl BinFile {
     }
 
     pub fn install_bin(&self) -> Result<(), BinstallError> {
-        // TODO: check if file already exists
+        if !self.source.try_exists()? {
+            return Err(BinstallError::BinFileNotFound(self.source.clone()));
+        }
+
         debug!(
             "Atomically install file from '{}' to '{}'",
             self.source.display(),
